requirements.txt
fastapi==0.115.6
google-ads==18.0.0
jinja2==3.1.2
mypy==1.15.0
psycopg2-binary==2.9.10
pydantic[email]==2.10.4
python-dotenv==1.0.1
requests==2.31.0
sqlalchemy==2.0.36
uvicorn==0.34.0